Reworked CityESP and Updated HoleUtil
Removed some left over debug lines.

Reworked CityESP to use damage calculations to accurately show which sides are cityable with immediate effect of dealing damage to the other player. Uses side, down and depth to find the area to search for a suitable damage location. Side works on each side, so with side being equal to 1 the actual width to search would be 3 (default 1 width).